Move dialogs to their own namespace
