+++ src/SerialLoops.Wpf/Program.cs
@@ -1,5 +1,6 @@
 ﻿using Eto.Wpf;
 using Eto.Wpf.Forms.ToolBar;
+using SerialLoops.Utility;
 using System;
 using System.Windows.Controls;
 using Application = Eto.Forms.Application;
